# src/docs/openapi-v1.yaml
openapi: 3.1.1
info:
  title: Odontofy API
  version: 1.0.0
  description: Contrato de la API v1 de Odontofy.
servers:
  - url: /api/v1
...
paths:
...
  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [Identity]
      summary: Rota la sesion refresh y emite un access token nuevo.
      security:
        - refreshCookie: []
      responses:
        '200':
          $ref: '#/components/responses/SessionCreated'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /auth/account-verification/confirm:
    post:
      operationId: confirmAccount
      tags: [Identity]
      summary: Consume un token de activacion de un solo uso.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountVerificationRequest'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/password/forgot:
    post:
      operationId: forgotPassword
      tags: [Identity]
      summary: Solicita recuperacion sin revelar si la cuenta existe.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/password/reset:
    post:
      operationId: resetPassword
      tags: [Identity]
      summary: Cambia la contrasena con un token de un solo uso.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/password/verify:
    post:
      operationId: verifyPassword
      tags: [Identity]
      summary: Reautentica al usuario para una operacion sensible.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyPasswordRequest'
      responses:
        '200':
          description: Contrasena verificada.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordVerifiedResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [Appointments]
      summary: Lista citas locales propias que intersectan un rango temporal.
      security:
        - bearerAuth: []
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { name: patientId, in: query, schema: { type: integer, minimum: 1 } }
        - { name: status, in: query, schema: { oneOf: [{ $ref: '#/components/schemas/AppointmentStatus' }, { const: all }], default: all } }
      responses:
        '200': { $ref: '#/components/responses/AppointmentList' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createAppointment
      tags: [Appointments]
      summary: Crea una cita local para un paciente propio.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAppointmentRequest' }
      responses:
        '201': { $ref: '#/components/responses/AppointmentResource' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/PatientNotFound' }
  /appointments/{appointmentId}:
    parameters:
      - { name: appointmentId, in: path, required: true, schema: { type: integer, minimum: 1, maximum: 4294967295 } }
    get:
      operationId: getAppointment
      tags: [Appointments]
      summary: Obtiene una cita local propia.
      security:
        - bearerAuth: []
      responses:
        '200': { $ref: '#/components/responses/AppointmentResource' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AppointmentNotFound' }
    patch:
      operationId: updateAppointment
      tags: [Appointments]
      summary: Actualiza campos permitidos de una cita local propia.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAppointmentRequest' }
      responses:
        '200': { $ref: '#/components/responses/AppointmentResource' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AppointmentNotFound' }
        '409': { $ref: '#/components/responses/AppointmentConflict' }
    delete:
      operationId: cancelAppointment
      tags: [Appointments]
      summary: Cancela logicamente una cita; es idempotente.
      security:
        - bearerAuth: []
      responses:
        '200': { $ref: '#/components/responses/AppointmentResource' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/AppointmentNotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: odontofy_refresh_v1
  responses:
    SessionCreated:
      description: Sesion autenticada.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Set-Cookie:
          description: Cookie refresh HttpOnly y SameSite Strict.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SessionResponse'
...
    EmptySuccess:
      description: Operacion completada.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EmptySuccessResponse'
    Accepted:
      description: Solicitud aceptada con respuesta anti-enumeracion.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EmptySuccessResponse'
    InvalidRequest:
      description: Payload o token de accion invalido.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Credenciales o sesion no validas.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Limite de solicitudes excedido.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    AppointmentResource:
      description: Cita local propia.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AppointmentResponse' }
    AppointmentList:
      description: Pagina de citas locales propias.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AppointmentListResponse' }
    AppointmentNotFound:
      description: La cita no existe o pertenece a otro usuario.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    AppointmentConflict:
      description: La cita esta cancelada o el calendario no esta disponible.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    PatientNotFound:
      description: El paciente no existe o pertenece a otro usuario.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  headers:
    RequestId:
      description: Identificador de correlacion de la solicitud.
      schema:
        type: string
        minLength: 1
        maxLength: 128
  schemas:
    BaseResponse:
      type: object
      required: [success, message, data, errors, requestId]
      properties:
        success:
          type: boolean
        message:
          type: string
        data: {}
        errors:
          oneOf:
            - type: array
              items:
                $ref: '#/components/schemas/ErrorItem'
            - type: 'null'
        requestId:
          type: string
        meta:
          type: object
          additionalProperties: true
    ErrorItem:
      type: object
      required: [code, details]
      properties:
        code:
          type: string
        details: {}
      additionalProperties: false
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              const: false
            data:
              type: 'null'
            errors:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/ErrorItem'
...
    EmailRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
      additionalProperties: false
    AccountVerificationRequest:
      type: object
      required: [userId, token]
      properties:
        userId:
          type: integer
          minimum: 1
        token:
          type: string
          pattern: '^[a-f0-9]{64}$'
      additionalProperties: false
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          pattern: '^[a-f0-9]{64}$'
        password:
          type: string
          minLength: 12
          maxLength: 72
      additionalProperties: false
    VerifyPasswordRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 72
      additionalProperties: false
...
    PublicUser:
      type: object
      required:
        - id
        - name
        - middleName
        - lastName
        - dateOfBirth
        - phone
        - avatar
        - email
        - showFinanceStats
        - isGoogleSynced
      properties:
        id:
          type: integer
        name:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          oneOf:
            - type: string
              format: date
            - type: 'null'
        phone:
          type: string
        avatar:
          type: string
        email:
          type: string
          format: email
        showFinanceStats:
          type: boolean
        isGoogleSynced:
          type: boolean
      additionalProperties: false
    SessionData:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/PublicUser'
      additionalProperties: false
    SessionResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              const: true
            data:
              $ref: '#/components/schemas/SessionData'
            errors:
              type: 'null'
...
    EmptySuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              const: true
            data:
              type: 'null'
            errors:
              type: 'null'
    PasswordVerifiedResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              const: true
            data:
              type: object
              required: [valid]
              properties:
                valid:
                  const: true
              additionalProperties: false
            errors:
              type: 'null'
...
    AppointmentStatus:
      type: string
      enum: [SCHEDULED, CONFIRMED, COMPLETED, CANCELLED, NO_SHOW]
    AppointmentSync:
      type: object
      required: [status, version, syncedAt, errorCode]
      properties:
        status:
          type: string
          enum: [NOT_CONNECTED, PENDING, SYNCED, FAILED]
        version: { type: integer, minimum: 1 }
        syncedAt:
          oneOf: [{ type: string, format: date-time }, { type: 'null' }]
        errorCode:
          oneOf: [{ type: string }, { type: 'null' }]
      additionalProperties: false
    AppointmentPatient:
      type: object
      required: [id, name, lastName]
      properties:
        id: { type: integer, minimum: 1 }
        name: { type: string }
        lastName: { type: string }
      additionalProperties: false
    Appointment:
      type: object
      required: [id, patientId, patient, startsAt, endsAt, timeZone, status, reason, note, cancelledAt, sync, createdAt, updatedAt]
      properties:
        id: { type: integer, minimum: 1 }
        patientId:
          oneOf: [{ type: integer, minimum: 1 }, { type: 'null' }]
        patient:
          oneOf: [{ $ref: '#/components/schemas/AppointmentPatient' }, { type: 'null' }]
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        timeZone: { type: string, maxLength: 64 }
        status: { $ref: '#/components/schemas/AppointmentStatus' }
        reason:
          oneOf: [{ type: string, maxLength: 255 }, { type: 'null' }]
        note:
          oneOf: [{ type: string, maxLength: 10000 }, { type: 'null' }]
        cancelledAt:
          oneOf: [{ type: string, format: date-time }, { type: 'null' }]
        sync: { $ref: '#/components/schemas/AppointmentSync' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      additionalProperties: false
    CreateAppointmentRequest:
      type: object
      required: [patientId, startsAt, endsAt, timeZone]
      properties:
        patientId: { type: integer, minimum: 1, maximum: 4294967295 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        timeZone: { type: string, maxLength: 64, example: America/Mexico_City }
        reason:
          oneOf: [{ type: string, maxLength: 255 }, { type: 'null' }]
        note:
          oneOf: [{ type: string, maxLength: 10000 }, { type: 'null' }]
      additionalProperties: false
    UpdateAppointmentRequest:
      type: object
      minProperties: 1
      properties:
        patientId: { type: integer, minimum: 1, maximum: 4294967295 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        timeZone: { type: string, maxLength: 64 }
        reason:
          oneOf: [{ type: string, maxLength: 255 }, { type: 'null' }]
        note:
          oneOf: [{ type: string, maxLength: 10000 }, { type: 'null' }]
      additionalProperties: false
...
    AppointmentResponse:
      allOf:
        - { $ref: '#/components/schemas/BaseResponse' }
        - type: object
          properties:
            success: { const: true }
            data: { $ref: '#/components/schemas/Appointment' }
            errors: { type: 'null' }
    AppointmentListResponse:
      allOf:
        - { $ref: '#/components/schemas/BaseResponse' }
        - type: object
          properties:
            success: { const: true }
            data:
              type: array
              items: { $ref: '#/components/schemas/Appointment' }
            errors: { type: 'null' }
...